Initial Upload
The code may have issues or bugs as it is the initial version. Reporting of issues or bugs or any other suggestions or ideas would be of great help. I admit this version lacks proper documentation. I am working on it and I will try to put a handy documentation in my next commit.